feat(ui): let color picker tool handle its events
